task5(personal_blog)/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==2.2.0
pillow==11.0.0
sqlparse==0.5.3
whitenoise==6.8.2
